+++ nequip/model/_build.py
@@ -61,6 +61,7 @@ def model_from_config(
             ), "inconsistant config & dataset"
         config["num_types"] = type_mapper.num_types
         config["type_names"] = type_mapper.type_names
+        config["type_to_chemical_symbol"] = type_mapper.type_to_chemical_symbol
 
     default_dtype = torch.get_default_dtype()
     model_dtype: torch.dtype = dtype_from_name(config.get("model_dtype", default_dtype))